+++ tools/firmware-utils/Makefile
@@ -39,8 +39,8 @@ define Host/Compile
 	$(call cc,encode_crc)
 	$(call cc,nand_ecc)
 	$(call cc,mkplanexfw sha1)
-	$(call cc,mktplinkfw md5, -Wall)
-	$(call cc,mktplinkfw2 md5)
+	$(call cc,mktplinkfw mktplinkfw-lib md5, -Wall -fgnu89-inline)
+	$(call cc,mktplinkfw2 mktplinkfw-lib md5, -fgnu89-inline)
 	$(call cc,tplink-safeloader md5, -Wall)
 	$(call cc,pc1crypt)
 	$(call cc,osbridge-crc)
