+++ tools/firmware-utils/src/mktplinkfw.c
@@ -28,19 +28,10 @@
 #include <netinet/in.h>
 
 #include "md5.h"
-
-#define ALIGN(x,a) ({ typeof(a) __a = (a); (((x) + __a - 1) & ~(__a - 1)); })
-#define ARRAY_SIZE(a) (sizeof((a)) / sizeof((a)[0]))
+#include "mktplinkfw-lib.h"
 
 #define HEADER_VERSION_V1	0x01000000
 #define HEADER_VERSION_V2	0x02000000
-
-#define MD5SUM_LEN	16
-
-struct file_info {
-	char		*file_name;	/* name of the file */
-	uint32_t	file_size;	/* length of the file */
-};
 
 struct fw_header {
 	uint32_t	version;	/* header version */
@@ -71,14 +62,6 @@ struct fw_header {
 	uint8_t		pad2[160];
 } __attribute__ ((packed));
 
-struct flash_layout {
-	char		*id;
-	uint32_t	fw_max_len;
-	uint32_t	kernel_la;
-	uint32_t	kernel_ep;
-	uint32_t	rootfs_ofs;
-};
-
 struct fw_region {
 	char		name[4];
 	uint32_t	code;
@@ -89,7 +72,7 @@ struct fw_region {
  * Globals
  */
 static char *ofname;
-static char *progname;
+char *progname;
 static char *vendor = "TP-LINK Technologies";
 static char *version = "ver. 1.0";
 static char *fw_ver = "0.0.0";
@@ -118,7 +101,6 @@ static struct file_info boot_info;
 static int combined;
 static int strip_padding;
 static int add_jffs2_eof;
-static unsigned char jffs2_eof_mark[4] = {0xde, 0xad, 0xc0, 0xde};
 static uint32_t fw_max_len;
 static uint32_t reserved_space;
 
@@ -190,42 +172,6 @@ static const struct fw_region regions[] = {
 	{"EU", 0},
 };
 
-/*
- * Message macros
- */
-#define ERR(fmt, ...) do { \
-	fflush(0); \
-	fprintf(stderr, "[%s] *** error: " fmt "\n", \
-			progname, ## __VA_ARGS__ ); \
-} while (0)
-
-#define ERRS(fmt, ...) do { \
-	int save = errno; \
-	fflush(0); \
-	fprintf(stderr, "[%s] *** error: " fmt ": %s\n", \
-			progname, ## __VA_ARGS__, strerror(save)); \
-} while (0)
-
-#define DBG(fmt, ...) do { \
-	fprintf(stderr, "[%s] " fmt "\n", progname, ## __VA_ARGS__ ); \
-} while (0)
-
-static struct flash_layout *find_layout(const char *id)
-{
-	struct flash_layout *ret;
-	struct flash_layout *l;
-
-	ret = NULL;
-	for (l = layouts; l->id != NULL; l++){
-		if (strcasecmp(id, l->id) == 0) {
-			ret = l;
-			break;
-		}
-	};
-
-	return ret;
-}
-
 static const struct fw_region * find_region(const char *country) {
 	size_t i;
 
@@ -269,59 +215,6 @@ static void usage(int status)
 	);
 
 	exit(status);
-}
-
-static void get_md5(const char *data, int size, uint8_t *md5)
-{
-	MD5_CTX ctx;
-
-	MD5_Init(&ctx);
-	MD5_Update(&ctx, data, size);
-	MD5_Final(md5, &ctx);
-}
-
-static int get_file_stat(struct file_info *fdata)
-{
-	struct stat st;
-	int res;
-
-	if (fdata->file_name == NULL)
-		return 0;
-
-	res = stat(fdata->file_name, &st);
-	if (res){
-		ERRS("stat failed on %s", fdata->file_name);
-		return res;
-	}
-
-	fdata->file_size = st.st_size;
-	return 0;
-}
-
-static int read_to_buf(const struct file_info *fdata, char *buf)
-{
-	FILE *f;
-	int ret = EXIT_FAILURE;
-
-	f = fopen(fdata->file_name, "r");
-	if (f == NULL) {
-		ERRS("could not open \"%s\" for reading", fdata->file_name);
-		goto out;
-	}
-
-	errno = 0;
-	fread(buf, fdata->file_size, 1, f);
-	if (errno != 0) {
-		ERRS("unable to read from file \"%s\"", fdata->file_name);
-		goto out_close;
-	}
-
-	ret = EXIT_SUCCESS;
-
- out_close:
-	fclose(f);
- out:
-	return ret;
 }
 
 static int check_options(void)
@@ -370,7 +263,7 @@ static int check_options(void)
 			return -1;
 		}
 	} else {
-		layout = find_layout(layout_id);
+		layout = find_layout(layouts, layout_id);
 		if (layout == NULL) {
 			ERR("unknown flash layout \"%s\"", layout_id);
 			return -1;
@@ -515,72 +408,6 @@ static void fill_header(char *buf, int len)
 		get_md5(buf, len, hdr->md5sum1);
 }
 
-static int pad_jffs2(char *buf, int currlen)
-{
-	int len;
-	uint32_t pad_mask;
-
-	len = currlen;
-	pad_mask = (64 * 1024);
-	while ((len < layout->fw_max_len) && (pad_mask != 0)) {
-		uint32_t mask;
-		int i;
-
-		for (i = 10; i < 32; i++) {
-			mask = 1 << i;
-			if (pad_mask & mask)
-				break;
-		}
-
-		len = ALIGN(len, mask);
-
-		for (i = 10; i < 32; i++) {
-			mask = 1 << i;
-			if ((len & (mask - 1)) == 0)
-				pad_mask &= ~mask;
-		}
-
-		for (i = 0; i < sizeof(jffs2_eof_mark); i++)
-			buf[len + i] = jffs2_eof_mark[i];
-
-		len += sizeof(jffs2_eof_mark);
-	}
-
-	return len;
-}
-
-static int write_fw(const char *data, int len)
-{
-	FILE *f;
-	int ret = EXIT_FAILURE;
-
-	f = fopen(ofname, "w");
-	if (f == NULL) {
-		ERRS("could not open \"%s\" for writing", ofname);
-		goto out;
-	}
-
-	errno = 0;
-	fwrite(data, len, 1, f);
-	if (errno) {
-		ERRS("unable to write output file");
-		goto out_flush;
-	}
-
-	DBG("firmware file \"%s\" completed", ofname);
-
-	ret = EXIT_SUCCESS;
-
- out_flush:
-	fflush(f);
-	fclose(f);
-	if (ret != EXIT_SUCCESS) {
-		unlink(ofname);
-	}
- out:
-	return ret;
-}
-
 static int build_fw(void)
 {
 	int buflen;
@@ -625,14 +452,14 @@ static int build_fw(void)
 			writelen = rootfs_ofs + rootfs_info.file_size;
 
 		if (add_jffs2_eof)
-			writelen = pad_jffs2(buf, writelen);
+			writelen = pad_jffs2(buf, writelen, layout->fw_max_len);
 	}
 
 	if (!strip_padding)
 		writelen = buflen;
 
 	fill_header(buf, writelen);
-	ret = write_fw(buf, writelen);
+	ret = write_fw(ofname, buf, writelen);
 	if (ret)
 		goto out_free_buf;
 
@@ -642,32 +469,6 @@ static int build_fw(void)
 	free(buf);
  out:
 	return ret;
-}
-
-/* Helper functions to inspect_fw() representing different output formats */
-static inline void inspect_fw_pstr(const char *label, const char *str)
-{
-	printf("%-23s: %s\n", label, str);
-}
-
-static inline void inspect_fw_phex(const char *label, uint32_t val)
-{
-	printf("%-23s: 0x%08x\n", label, val);
-}
-
-static inline void inspect_fw_phexdec(const char *label, uint32_t val)
-{
-	printf("%-23s: 0x%08x / %8u bytes\n", label, val, val);
-}
-
-static inline void inspect_fw_pmd5sum(const char *label, const uint8_t *val, const char *text)
-{
-	int i;
-
-	printf("%-23s:", label);
-	for (i=0; i<MD5SUM_LEN; i++)
-		printf(" %02x", val[i]);
-	printf(" %s\n", text);
 }
 
 static int inspect_fw(void)
